Refactorings

+++ OctreeApi/OctreeLibrary/src/include/tools.cuh
@@ -1,26 +1,23 @@
 #pragma once
 
 
+#include "metadata.cuh"
+#include "types.cuh"
 #include <cuda_runtime_api.h>
 #include <memory>
-
-#include "octree_metadata.h"
-#include "types.cuh"
-
 
 using namespace std;
 
 
 namespace tools {
 
-uint32_t getOctreeLevel (uint32_t gridSize);
+uint8_t getOctreeLevel (uint32_t gridSize);
 uint32_t getOctreeGrid (uint32_t octreeLevel);
 uint32_t getNodeAmount (uint32_t octreeLevel);
 uint32_t getNodeOffset (uint32_t octreeLevel, uint32_t octreeDepth);
 
 template <typename coordinateType>
-unique_ptr<CudaArray<uint8_t>> generate_point_cloud_cuboid (
-        uint32_t sideLength, PointCloudMetadata& metadata);
+unique_ptr<CudaArray<uint8_t>> generate_point_cloud_cuboid (uint32_t sideLength, PointCloudMetadata& metadata);
 void printKernelDimensions (dim3 block, dim3 grid);
 void create1DKernel (dim3& block, dim3& grid, uint32_t pointCount);
 
@@ -30,8 +27,8 @@ __host__ __device__ void mapFromDenseIdxToDenseCoordinates (
 
 // See OctreeConverter : chunker_countsort_laszip.cpp :131
 template <typename coordinateType>
-__device__ uint32_t calculateGridIndex (
-        const Vector3<coordinateType>* point, PointCloudMetadata const& metadata, uint32_t gridSize)
+__device__ uint32_t
+        calculateGridIndex (const Vector3<coordinateType>* point, PointCloudMetadata const& metadata, uint32_t gridSize)
 {
     double sizeX = metadata.bbCubic.max.x - metadata.bbCubic.min.x;
     double sizeY = metadata.bbCubic.max.y - metadata.bbCubic.min.y;
